# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Enquiry {
  id: number;
  name: string;
  partnerName: string;
  email: string;
  phone: string;
  weddingDate?: string | null;
  venue: string;
  service: string;
  howHeard?: string | null;
  details: string;
  status: string;
  createdAt: string;
}

export interface CreateEnquiry {
  name: string;
  partnerName: string;
  email: string;
  phone: string;
  weddingDate?: string | null;
  venue: string;
  service: string;
  howHeard?: string | null;
  details: string;
}

export type UpdateEnquiryStatusStatus = typeof UpdateEnquiryStatusStatus[keyof typeof UpdateEnquiryStatusStatus];


export const UpdateEnquiryStatusStatus = {
  new: 'new',
  read: 'read',
  replied: 'replied',
} as const;

export interface UpdateEnquiryStatus {
  status: UpdateEnquiryStatusStatus;
}

export interface ErrorResponse {
  error: string;
}