# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: enquiries
    description: Wedding enquiry form submissions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /enquiries:
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries
      description: Returns all form submissions, newest first
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a new enquiry
      description: Saves a contact form submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiry"
      responses:
        "201":
          description: Enquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /enquiries/{id}:
    get:
      operationId: getEnquiry
      tags: [enquiries]
      summary: Get a single enquiry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enquiry details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEnquiryStatus
      tags: [enquiries]
      summary: Update enquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnquiryStatus"
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Enquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        partnerName:
          type: string
        email:
          type: string
        phone:
          type: string
        weddingDate:
          type: string
          nullable: true
        venue:
          type: string
        service:
          type: string
        howHeard:
          type: string
          nullable: true
        details:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - partnerName
        - email
        - phone
        - venue
        - service
        - details
        - status
        - createdAt

    CreateEnquiry:
      type: object
      properties:
        name:
          type: string
        partnerName:
          type: string
        email:
          type: string
        phone:
          type: string
        weddingDate:
          type: string
          nullable: true
        venue:
          type: string
        service:
          type: string
        howHeard:
          type: string
          nullable: true
        details:
          type: string
      required:
        - name
        - partnerName
        - email
        - phone
        - venue
        - service
        - details

    UpdateEnquiryStatus:
      type: object
      properties:
        status:
          type: string
          enum: [new, read, replied]
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error