# lib/api-zod/src/generated/types/createEnquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateEnquiry {
  name: string;
  partnerName: string;
  email: string;
  phone: string;
  weddingDate?: string | null;
  venue: string;
  service: string;
  howHeard?: string | null;
  details: string;
}