# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Enquiry {
  id: number;
  name: string;
  partnerName: string;
  email: string;
  phone: string;
  weddingDate?: string | null;
  venue: string;
  service: string;
  howHeard?: string | null;
  details: string;
  status: string;
  createdAt: string;
}