# lib/db/src/schema/enquiries.ts
import { pgTable, text, serial, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const enquiriesTable = pgTable("enquiries", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  partnerName: text("partner_name").notNull(),
  email: text("email").notNull(),
  phone: text("phone").notNull(),
  weddingDate: text("wedding_date"),
  venue: text("venue").notNull(),
  service: text("service").notNull(),
  howHeard: text("how_heard"),
  details: text("details").notNull(),
  status: text("status").notNull().default("new"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertEnquirySchema = createInsertSchema(enquiriesTable).omit({
  id: true,
  createdAt: true,
  status: true,
});

export type InsertEnquiry = z.infer<typeof insertEnquirySchema>;
export type Enquiry = typeof enquiriesTable.$inferSelect;